F I N A L L Y

+++ boards/platform/stm32f4/src/bsp_os.cpp
@@ -49,7 +49,7 @@ namespace bsp {
         htim_os = htim;
     }
 
-    uint64_t GetHighresTickMicroSec(void) {
+    uint16_t GetHighresTickMicroSec(void) {
         return getRunTimeCounterValue();
     }
     uint32_t GetHighresTickMilliSec(void) {
